requirements.txt
# =============================================================================
# PICKLE — 신대방삼거리 맛집 추천 챗봇 의존성
# Python 3.10+
# =============================================================================

# -----------------------------------------------------------------------------
# 환경 설정 (.env 로딩)
#   - src/config.py, database/sql/utils.py, main.py, frontend/app.py
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0

# -----------------------------------------------------------------------------
# LLM / Embedding (OpenAI)
#   - src/llm_client.py, database/sql/utils.py, embedding_*.ipynb
# -----------------------------------------------------------------------------
openai>=1.40.0

# -----------------------------------------------------------------------------
# LangChain 계열 (LangGraph 파이프라인 + ChatOpenAI / OpenAIEmbeddings)
#   - src/router.py, src/generator.py, src/embeddings.py
#   - src/slot_extractor.py (langchain_core.tools.tool)
#   - src/pipeline.py      (langgraph StateGraph)
# -----------------------------------------------------------------------------
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# -----------------------------------------------------------------------------
# 데이터 처리 / 수치 연산
#   - database/sql/utils.py (임베딩 base64 디코딩 + 코사인 유사도)
#   - 전처리 노트북 전반 (pandas / numpy)
# -----------------------------------------------------------------------------
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0

# -----------------------------------------------------------------------------
# Frontend (Streamlit + Kakao Map iframe)
#   - main.py, frontend/app.py
# -----------------------------------------------------------------------------
streamlit>=1.36.0

# -----------------------------------------------------------------------------
# 크롤링 / 파싱 (DB 재구축 시 필요)
#   - database/raw/0_page_search_parser.py          (BeautifulSoup)
#   - database/raw/1_dynamic_crawling.ipynb         (Selenium)
#   - database/processed/1_html_to_csv.ipynb        (BeautifulSoup)
# -----------------------------------------------------------------------------
beautifulsoup4>=4.12.0
lxml>=5.0.0
selenium>=4.20.0

# -----------------------------------------------------------------------------
# Notebook 실행 보조
#   - embedding_review.ipynb 등에서 tqdm 진행률 표시
#   - 주피터로 database/*.ipynb 를 열어야 할 때 필요
# -----------------------------------------------------------------------------
tqdm>=4.66.0
ipykernel>=6.29.0
jupyter>=1.0.0
